requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=0.9.0=pypi_0
attrs=19.3.0=py_0
backcall=0.1.0=py37_0
blas=1.0=mkl
bleach=3.1.0=py37_0
ca-certificates=2020.7.22=0
cachetools=4.1.1=pypi_0
certifi=2020.6.20=py37_0
cffi=1.13.2=py37h7a1dbc1_0
chardet=3.0.4=pypi_0
colorama=0.4.1=py37_0
cudatoolkit=10.1.168=0
cycler=0.10.0=py37_0
decorator=4.4.1=py_0
defusedxml=0.6.0=py_0
entrypoints=0.3=py37_0
freetype=2.9.1=ha9979f8_1
google-auth=1.20.1=pypi_0
google-auth-oauthlib=0.4.1=pypi_0
grpcio=1.31.0=pypi_0
h5py=2.10.0=py37h5e291fa_0
hdf5=1.10.4=h7ebc959_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha66f8fd_1
idna=2.10=pypi_0
importlib_metadata=0.23=py37_0
intel-openmp=2019.4=245
ipykernel=5.1.3=py37h39e3cac_0
ipython=7.9.0=py37h39e3cac_0
ipython_genutils=0.2.0=py37_0
ipywidgets=7.5.1=py_0
jedi=0.15.1=py37_0
jinja2=2.10.3=py_0
joblib=0.15.1=py_0
jpeg=9b=hb83a4c4_2
jsonschema=3.2.0=py37_0
jupyter=1.0.0=py37_7
jupyter_client=5.3.4=py37_0
jupyter_console=6.0.0=py37_0
jupyter_core=4.6.1=py37_0
kiwisolver=1.1.0=py37ha925a31_0
laspy=1.7.0=pypi_0
libpng=1.6.37=h2a8f88b_0
libsodium=1.0.16=h9d3ae62_0
libtiff=4.1.0=h56a325e_0
llvmlite=0.29.0=py37ha925a31_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.2.2=pypi_0
markupsafe=1.1.1=py37he774522_0
matplotlib=3.1.0=py37hc8f65d3_0
mistune=0.8.4=py37he774522_0
mkl=2019.4=245
mkl-service=2.3.0=py37hb782905_0
mkl_fft=1.0.15=py37h14836fe_0
mkl_random=1.1.0=py37h675688f_0
more-itertools=7.2.0=py37_0
msys2-conda-epoch=20160418=1
nbconvert=5.6.1=py37_0
nbformat=4.4.0=py37_0
ninja=1.9.0=py37h74a9793_0
notebook=6.0.2=py37_0
numba=0.45.1=py37hf9181ef_0
numpy=1.17.3=py37h4ceb530_0
numpy-base=1.17.3=py37hc3f5095_0
oauthlib=3.1.0=pypi_0
olefile=0.46=py37_0
open3d=0.10.0.1=py37_0
openssl=1.1.1g=he774522_1
pandoc=2.2.3.2=0
pandocfilters=1.4.2=py37_1
parso=0.5.1=py_0
pickleshare=0.7.5=py37_0
pillow=6.2.1=py37hdc69c19_0
pip=19.3.1=py37_0
prometheus_client=0.7.1=py_0
prompt_toolkit=2.0.10=py_0
protobuf=3.13.0=pypi_0
py7zr=0.9.5=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.19=py37_0
pycryptodome=3.9.8=pypi_0
pygments=2.4.2=py_0
pyparsing=2.4.5=py_0
pyqt=5.9.2=py37h6538335_2
pyreadline=2.1=py37_1
pyrsistent=0.15.5=py37he774522_0
python=3.7.4=h5263a28_0
python-dateutil=2.8.1=py_0
pytorch=1.3.0=py3.7_cuda101_cudnn7_0
pytz=2019.3=py_0
pywin32=223=py37hfa6e2cd_1
pywinpty=0.5.5=py37_1000
pyzmq=18.1.0=py37ha925a31_0
qt=5.9.7=vc14h73c81de_0
qtconsole=4.6.0=py_0
requests=2.24.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.6=pypi_0
scikit-learn=0.23.1=py37h25d0782_0
scipy=1.5.0=py37h9439919_0
send2trash=1.5.0=py37_0
setuptools=41.6.0=py37_0
sip=4.19.8=py37h6538335_0
six=1.13.0=py37_0
sqlite=3.30.1=he774522_0
tbb=2019.4=h74a9793_0
tensorboard=2.3.0=pypi_0
tensorboard-plugin-wit=1.7.0=pypi_0
terminado=0.8.3=py37_0
testpath=0.4.4=py_0
texttable=1.6.2=pypi_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.8=hfa6e2cd_0
torchvision=0.4.1=py37_cu101
tornado=6.0.3=py37he774522_0
tqdm=4.48.2=py_0
traitlets=4.3.3=py37_0
urllib3=1.25.10=pypi_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_0
wcwidth=0.1.7=py37_0
webencodings=0.5.1=py37_1
werkzeug=1.0.1=pypi_0
wheel=0.33.6=py37_0
widgetsnbextension=3.5.1=py37_0
wincertstore=0.2=py37_0
winpty=0.4.3=4
xz=5.2.4=h2fa13f4_4
zeromq=4.3.1=h33f27b4_3
zipp=0.6.0=py_0
zlib=1.2.11=h62dcd97_3
zstd=1.3.7=h508b16e_0
